Refactor the Dump command, start on the Generate command
