+++ rbx_reflector/Cargo.toml
@@ -6,7 +6,8 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-rbx_dom_weak = { version = "2.3.0", path = "../rbx_dom_weak" }
+rbx_types = { version = "1.3.0", path = "../rbx_types", features = ["serde"] }
+rbx_reflection = { version = "4.2.0", path = "../rbx_reflection" }
 
 anyhow = "1.0.57"
 clap = { version = "3.1.12", features = ["derive"] }
@@ -14,7 +15,6 @@ roblox_install = "1.0.0"
 serde = { version = "1.0.136", features = ["derive"] }
 serde_json = "1.0.79"
 tempfile = "3.3.0"
-rbx_reflection = "4.2.0"
 rmp-serde = "1.1.0"
 fs-err = "2.7.0"
 notify = "5.0.0"
